#9675
Cast the width argument of every "%.*s" to int at the call site, in 39 places
across 10 files, plus one "%d" that was being passed a pid_t.

The "*" conversion takes an int.  Passing a uint32_t is undefined behavior
even where it happens to work, and it works on the linux hosts only because
int and uint32_t are the same width and are passed identically there, and
because modern gcc does not warn.  gcc 2.95.2 on solaris8sparc does warn, and
-Werror turns each one into a build failure:

	cc1plus: warnings being treated as errors
	sqlrclient.cpp:1462: warning: field width is not type int (arg 3)
	gmake[2]: *** [sqlrclient.lo] Error 1

The sites, all found by one "gmake -k" pass so the compiler named every one
rather than surfacing them a build at a time:

	src/protocols/firebird.cpp          6
	src/protocols/mysql.cpp             1
	src/protocols/postgresql.cpp        4
	src/protocols/sqlrclient.cpp        8
	src/protocols/tds.cpp               3
	src/protocols/teradata.cpp          9
	src/triggers/replay.cpp             5
	src/triggers/upsert.cpp             2
	src/triggers/splitmultiinsert.cpp   1
	src/triggers/globaltemptables.cpp   1

The width is not always the first argument after the format, which is worth
knowing before anyone tries to automate a change like this: in
"%s(%.*s(%d)) = " it is the second, and teradata.cpp:4655
("raw query: (%d) %.*s") had two wrong arguments on one call - the pid for the
%d and the length for the %.*s.  gcc's "(arg N)" is what identifies the right
one, counting this as arg 1 and the format as arg 2.

Where an argument already carried a wider cast, that cast was replaced rather
than nested, so nothing reads "(int)(int32_t)".

Verified on solaris8sparc: all 10 files recompile with zero warnings and zero
errors under -Wall -Wformat-nonliteral -Wformat-security -Werror with gcc
2.95.2, and a full "gmake -k" over src/ now reports no diagnostics at all
where it previously reported 40.

+++ src/protocols/firebird.cpp
@@ -6478,7 +6478,7 @@ bool sqlrprotocol_firebird::prepareOrExecImmediate(bool execimmediate) {
 	debugWrite("statement handle: %u",stmthandle);
 	debugWrite("dialect: %u",dialect);
 	debugWrite("query length: %u",querylen);
-	debugWrite("query: \"%.*s\"",querylen,(query)?query:"");
+	debugWrite("query: \"%.*s\"",(int)querylen,(query)?query:"");
 	debugWrite("requested sql info items length: %u",itemslen);
 	debugWrite("response buffer length: %u",respbufferlen);
 
@@ -6512,7 +6512,7 @@ bool sqlrprotocol_firebird::runPreparedQuery(bool execimmediate,
 	debugWrite("in transaction: %s",(intransaction)?"yes":"no");
 	debugWrite("read only: %s",(trreadonly)?"yes":"no");
 	debugWrite("query length: %u",querylen);
-	debugWrite("query: \"%.*s\"",querylen,(query)?query:"");
+	debugWrite("query: \"%.*s\"",(int)querylen,(query)?query:"");
 	debugWrite("requested sql info items length: %u",itemslen);
 
 	// bounds check, before the query can size anything
@@ -7167,7 +7167,7 @@ bool sqlrprotocol_firebird::runOnCursor(sqlrservercursor *cursor,
 	debugStart("run on cursor");
 
 	debugWrite("cursor id: %u",cont->getId(cursor));
-	debugWrite("query: \"%.*s\"",querylen,query);
+	debugWrite("query: \"%.*s\"",(int)querylen,query);
 	debugWrite("query length: %u",querylen);
 	debugWrite("output field count: %u",outfieldcount);
 
@@ -10486,7 +10486,7 @@ bool sqlrprotocol_firebird::execBatchMessage(sqlrservercursor *cursor,
 		err->sqlstate[sizeof(err->sqlstate)-1]='\0';
 
 		debugWrite("execute failed");
-		debugWrite("error: \"%.*s\"",(uint32_t)errorsize,errorstring);
+		debugWrite("error: \"%.*s\"",(int)errorsize,errorstring);
 		debugWrite("error number: %lld",(long long)errnum);
 		debugWrite("sqlcode: %d",(int)sqlcode);
 		debugWrite("live connection: %s",(liveconnection)?"yes":"no");
@@ -11443,7 +11443,7 @@ bool sqlrprotocol_firebird::sendCursorError(const char *title,
 
 	debugStart("send cursor error");
 	debugWrite("preparing: %s",(preparing)?"yes":"no");
-	debugWrite("error: \"%.*s\"",(uint32_t)errorsize,errorstring);
+	debugWrite("error: \"%.*s\"",(int)errorsize,errorstring);
 	debugWrite("error size: %u",errorsize);
 	debugWrite("error number: %lld",(long long)errnum);
 	debugWrite("live connection: %s",(liveconnection)?"yes":"no");
@@ -11663,7 +11663,7 @@ bool sqlrprotocol_firebird::buildBindProbe(sqlrservercursor *cursor,
 
 	*bindcount=cont->countBindVariables(query,querylen);
 
-	debugWrite("query: \"%.*s\"",querylen,query);
+	debugWrite("query: \"%.*s\"",(int)querylen,query);
 	debugWrite("bind count: %hd",*bindcount);
 
 	if (!*bindcount) {
